Shift paradigm from layers to nodes
This commit shifts the central paradigm of nin from being a layer-based
system to being a node-based system. Not all features have been ported
over to the new system, although it is certainly complete enough to be
used to make demos. Hopefully we are willing to merge this sooner rather
than later, and we can then later fix the rest of the features as they
are needed.

* Overview

Before, a demo in nin would consist of several layers in an ordered
list. Each layer would have a start time and end time, and each
layer's update and render methods would be called with the usual
update/render call semantics while demo time was within a layer's start
and end times. Each layer exposed an EffectComposer pass, and all passes
for active layers were chained together in a top-to-bottom such that the
next layer could optionally access the rendered data of the previous
layer.

Now, a demo in nin consists of several nodes in a graph. Nodes do not
have start times or end times. Rather, the graph is traversed from a
designated root node in order call update and render for each of the
reachable nodes. A node may enable or disable their inputs, which serves
as temporary graph modification in order to "connect and disconnect"
different nodes at different times in the demo. This is the spiritual
replacement of start times and end times from the layer system.  The big
advantage of the new node-based system over the old layer system  is
that it is now easier to create more advanced multi-pass rendering
pipelines.

* Nodes and the graph

Nodes have different types, and all nodes inherit from the base
NIN.Node. A node type may declare a number of named and typed inputs and
outputs. Nodes may produce data and store it in their outputs. Connected
nodes may then read that same data from their inputs. The directed graph
of connections from node outputs to node inputs may not contain cycles.

The graph is stored in res/graph.json, which is analogous to the old
res/layers.json. Because of the frequency of graph edits in the UI
(particularly dragging nodes to change their position), and in the
interest of simplicity and performance, res/graph.json is not actively
watched for changes on disk after the first load at the beginning of
`nin run`. The intention is that most changes to res/graph.json should
be made via the UI.

Currently, all graphs *must* have a node of type NIN.RootNode called
root in order for the demo to update and render properly. This node is
the root node that is used for traversal during update and render, and
the texture made available to the root node's only input root.screen is
what is finally showed on the screen.

The LayerManager of old has been replaced with an analogous NodeManager.

* Current state of features

This section elaborates upon the different features of nin and their
current state with regards to the paradigm shift. Features that are not
affected are probably not mentioned here.

** Node graph viewer / editor

Replacing the old layer viewer/editor, the node graph viewer / editor
currently supports visualization of the graph including connections and
activeness of nodes. Nodes are persistently draggable. The viewer is
pan/zoom-enabled, but has no pan/zoom persistence. It is currently not
possible to create or delete nodes or connections.

** Generate

This has not been touched, so it most likely does not work.

** Compilation

This has not been touched, so it most likely does not work.

** Init

This has not been touched, so it most likely does not work.

** Backend res/layers.json / res/graph.js manipulation

This has been updated and works for whatever it is being used for in
this commit, but has not been scrutinized for completeness.

** Shader updates

This is probably broken. Perhaps we can leverage the node system in
order to make it less grep-y in the future.

** Camera paths, fly-around mode

This is probably broken.

** Layer / Node inspection

Nodes can no longer be inspected, but inputs and outputs can. Currently,
the only inspectable inputs/outputs are texture input/outputs. When
inspected, they will render their texture to the demo screen after the
regular render pass is finished.